Update and rename purge.py to msg_deleting.py
